+++ src/main/java/com/oms/dto/responses/PODetailAsList.java
@@ -8,18 +8,18 @@ import java.util.Date;
 
 @Data
 public class PODetailAsList {
-private Long id;
+    private Long id;
     private String poDate;
     private String poNumber;
     private Long customerId;
     private String customerName;
     private String orderStatus;
-    private float totalAmount;
+    private double totalAmount;
 
     public PODetailAsList(){
 
     }
-    public PODetailAsList(Long id,String poNumber,Date poDate,String orderStatus,float totalAmount,Long customerId,String customerName)
+    public PODetailAsList(Long id,String poNumber,Date poDate,String orderStatus,double totalAmount,Long customerId,String customerName)
     {
         this.id=id;
         this.poDate=new SimpleDateFormat("dd/MM/yyyy").format(poDate);
